+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Replace `NavigationNodeFactory.Companion.Typed` factory into `NavigationNodeFactory.Typed`
   * Change order of `NavigationNodeFactory.Typed` type args
   * Add default aggregator for `NavigationNodeFactory`
+  * Add extension `changesInSubtreeFlow` for either chains and nodes
 
 ## 0.5.6